+++ object-storage-spring-boot-autoconfigure/src/main/java/com/github/anilople/object/storage/autoconfigure/ObjectStorageAutoConfiguration.java
@@ -19,7 +19,7 @@
 package com.github.anilople.object.storage.autoconfigure;
 
 import com.github.anilople.object.storage.autoconfigure.constant.ObjectStorageConstants;
-import com.github.anilople.object.storage.core.ObjectStorageEndpoint;
+import com.github.anilople.object.storage.core.config.ObjectStorageEndpoint;
 import java.net.URI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
